requirement.txt
Flask==2.3.3
tensorflow==2.13.0
Keras==2.13.1
numpy==1.24.3
Pillow==9.5.0
matplotlib==3.7.1
scikit-learn==1.3.2
